+++ ICD.Connect.Protocol.Network/Settings/UriPropertiesConsole.cs
@@ -1,0 +1,57 @@
+using System;
+using System.Collections.Generic;
+using ICD.Common.Utils;
+using ICD.Connect.API.Commands;
+using ICD.Connect.API.Nodes;
+
+namespace ICD.Connect.Protocol.Network.Settings
+{
+	public static class UriPropertiesConsole
+	{
+		/// <summary>
+		/// Gets the child console nodes.
+		/// </summary>
+		/// <param name="instance"></param>
+		/// <returns></returns>
+		public static IEnumerable<IConsoleNodeBase> GetConsoleNodes(IUriProperties instance)
+		{
+			if (instance == null)
+				throw new ArgumentNullException("instance");
+
+			yield break;
+		}
+
+		/// <summary>
+		/// Calls the delegate for each console status item.
+		/// </summary>
+		/// <param name="instance"></param>
+		/// <param name="addRow"></param>
+		public static void BuildConsoleStatus(IUriProperties instance, AddStatusRowDelegate addRow)
+		{
+			if (instance == null)
+				throw new ArgumentNullException("instance");
+
+			addRow("Username", instance.UriUsername);
+			addRow("Password", StringUtils.PasswordFormat(instance.UriPassword));
+			addRow("Host", instance.UriHost);
+			addRow("Port", instance.UriPort);
+			addRow("Scheme", instance.UriScheme);
+			addRow("Path", instance.UriPath);
+			addRow("Query", instance.UriQuery);
+			addRow("Fragment", instance.UriFragment);
+		}
+
+		/// <summary>
+		/// Gets the child console commands.
+		/// </summary>
+		/// <param name="instance"></param>
+		/// <returns></returns>
+		public static IEnumerable<IConsoleCommand> GetConsoleCommands(IUriProperties instance)
+		{
+			if (instance == null)
+				throw new ArgumentNullException("instance");
+
+			yield break;
+		}
+	}
+}